+++ layouts/UserProfileLayout.tsx
@@ -1,0 +1,25 @@
+//components
+import Header from "../containers/Header";
+import Footer from "../containers/Footer";
+import SideBarMenu from "../components/SideBarMenu";
+
+//interface
+import { BaseLayoutInterface } from "../types/Layout.types";
+
+
+function UserProfileLayout({ children }: BaseLayoutInterface):JSX.Element {
+
+
+  return (
+    <div className="page-content">
+      <Header/>
+      <div className="base-page-layout user-profile-layout">
+        <SideBarMenu/>
+        { children }
+      </div>
+      <Footer/>
+    </div>
+  );
+}
+
+export default UserProfileLayout;