+++ types/Elements.types.ts
@@ -1,3 +1,5 @@
+import { ReactNode } from "react";
+
 export interface BaseButtonProps {
   onClick: () => void;
   size?: "large" | "middle" | "small"
@@ -7,4 +9,14 @@ export interface BaseButtonProps {
 export interface ButtonProps {
 	className?: string;
 	label?: string;
+}
+
+export interface LinkCustomInterface {
+  href: string,
+  as?: string,
+  className?: string,
+  onCLick?: () => void,
+  label?: string | number,
+  size?: "s" | "l",
+  children?: ReactNode | Array<ReactNode>;
 }