+++ types/Layout.types.ts
@@ -2,10 +2,14 @@ import { ReactNode } from "react";
 
 
 export interface BaseLayoutInterface {
-	children: ReactNode;
+	children: ReactNode | Array<ReactNode>;
 }
 
 export interface ProductsLayoutInterface {
 	children: ReactNode;
+}
+
+export interface AuthLayoutInterface extends BaseLayoutInterface {
+  formTitle?: string;
 }